src/requirements.txt
fastapi==0.75.1
uvicorn==0.15.0
selenium
python-dotenv
